Mudança status slack

+++ slack_hafutech/hafutech/src/main/java/sptech/school/hafutech/repository/EmpresaRepository.java
@@ -8,6 +8,6 @@ import sptech.school.hafutech.entity.Empresa;
 @Repository
 public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
     @Query("SELECT id FROM Empresa ORDER BY id DESC LIMIT 1")
-    Long pegarUltimoId();
+    Integer pegarUltimoId();
 
 }
